fix: user creation should happen on connect
