+++ docker-compose.yml
@@ -20,6 +20,7 @@ services:
       - ./prisma:/usr/src/app/prisma
       - ./graphql/src:/usr/src/app/graphql/src
       - ./graphql/index.ts:/usr/src/app/graphql/index.ts
+      - ./lib:/usr/src/app/lib
     ports:
       - 8080:4000
     command: [bash, -c, "cd graphql && npm start"]
