+++ lib/auth/index.ts
@@ -1,3 +1,2 @@
 export { AuthContext } from '../auth/AuthContext';
-export { IUserDataContext } from '../auth/UserDataContext';
 export { IDataContext } from '../auth/types';
