+++ lib/auth/types.ts
@@ -1,15 +1,18 @@
-import { Prisma } from '@prisma/client';
+import { Prisma, User } from '@prisma/client';
 
 export interface IDataContext<T> {
-  getAll?: () => Promise<T[]>;
+  getAll: () => Promise<T[]>;
   get(id: string): Promise<T | null>;
   post(item: T): Promise<T>;
-  // delete(id: string): Promise<T>;
-  // put(item: T): Promise<T>;
+}
+
+export interface IUserDataContext extends IDataContext<User> {
+  getByEmail(email: string): Promise<User | null>;
+  createUser(email?: string): Promise<User | null>;
 }
 
 const userWithSessions = Prisma.validator<Prisma.UserArgs>()({
   include: { sessions: true },
 });
 
-export type UserWithSesions = Prisma.UserGetPayload<typeof userWithSessions>;
+export type UserWithSessions = Prisma.UserGetPayload<typeof userWithSessions>;
